# chore: Minor rbac memory optimization