--- a/coderd/rbac/authz_test.go
+++ b/coderd/rbac/authz_test.go
@@ -97,12 +97,14 @@ func benchmarkUserCases() (cases []benchmarkCase, users uuid.UUID, orgs []uuid.U
 			},
 		},
 		{
+			// This test should only use static roles. AKA no org roles.
 			Name: "StaticRoles",
 			Actor: rbac.Subject{
 				// Give some extra roles that an admin might have
 				Roles: rbac.RoleNames{
 					"auditor", rbac.RoleMember(), rbac.RoleMember(),
-					rbac.RoleTemplateAdmin(), rbac.RoleUserAdmin()},
+					rbac.RoleTemplateAdmin(), rbac.RoleUserAdmin(),
+				},
 				ID:     user.String(),
 				Scope:  rbac.ScopeAll,
 				Groups: noiseGroups,